requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
Flask==2.2.2
fonttools==4.37.4
importlib-metadata==5.0.0
ipykernel==6.15.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
pyzmq==23.2.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sklearn==0.0
stack-data==0.5.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.8.1
gunicorn
streamlit